feat: make refunds work — persist them, and stop refunding twice
processRefund() called the driver, returned a DTO and wrote nothing.
`$transaction->refunds()->sum('amount')` — the expression a "can this still
be refunded?" guard is naturally built on — therefore always returned zero,
so any such guard passed every time and the same transaction could be
refunded repeatedly. Models\Refund could not have helped: it still described
the 1.x schema (processor_refund_id/processor_response against a migration
writing provider_refund_id/provider_payload, a uuid key against a bigint
one, an integer cast on a decimal column) and nothing ever wrote through it,
so the mismatch stayed invisible.

Refund amounts widen from int to float across the interface, both DTOs, the
logger and the fake. Every bundled driver already cast the int straight to
float before sending it, so the wire format is unchanged — but the parameter
truncated on the way in, and a $95.50 refund reached the PSP as $95. Units
are major, matching transactions.amount. TransactionWebhookUpdate::$amount
goes with them: as an int it truncated a webhook-reported 95.50 to 95 before
the OnHold tolerance comparison ran against a decimal column.

processRefund() now reserves before it calls out: lock the transaction, sum
refunds in succeeded|pending|processing, refuse anything exceeding the
remainder, insert a pending row, release the lock, call the PSP, write the
outcome back. Reserving under the lock is what stops two concurrent requests
seeing the same balance; calling the PSP outside it is what stops a slow
gateway blocking every webhook for that transaction. A refusal or a thrown
exception marks the row failed and frees the balance again — otherwise an
outage would permanently consume a customer's refundable amount.

RefundSucceeded/RefundFailed carry the Refund and are finally dispatched.

+++ src/Abstracts/AbstractPaymentProcessor.php
@@ -25,9 +25,9 @@ abstract class AbstractPaymentProcessor implements PaymentProcessorInterface
 
     abstract public function charge(array $data): PaymentResult;
 
-    abstract public function refund(string $transactionId, ?int $amount = null): RefundResult;
+    abstract public function refund(string $transactionId, ?float $amount = null): RefundResult;
 
-    public function capture(string $transactionId, ?int $amount = null): PaymentResult
+    public function capture(string $transactionId, ?float $amount = null): PaymentResult
     {
         throw new \BadMethodCallException('Capture method not implemented for ' . $this->getName());
     }
